:penguin:

+++ RIProject/src/main/java/RI/TF_IDF.java
@@ -29,7 +29,7 @@ public class TF_IDF {
 
     public void calculateIDF(HashMap<String, Tuple> invertedIndex, int nFiles) {
         for (String term : invertedIndex.keySet()) {
-            invertedIndex.get(term).setIDF(Math.log((double) nFiles / invertedIndex.get(term).getWeightDoc().size()));
+            invertedIndex.get(term).setIDF(Math.log(nFiles / invertedIndex.get(term).getWeightDoc().size()));
         }
     }
